Add link to documentation
So that documentation is easily available and only needs to be maintained in one place, the README for this component should feature a link to the proper component doc published in the App Store Guide. I have already added the link for the doc to the component’s **Documentation** tab in the App Store. Please contact me with any questions.