Remove double encoding
When passing a parameter that has a + in it the end result of this call is a blank. 
E.g. 't%2BprkqguXi' becomes 't BprkqguXi' when it should be 't+BprkqguXi'

It is because the parameter is decoded twice, the second time a '+' becomes a blank. This change is just removing the extra decode call. 

https://forum.mendixcloud.com/link/questions/99153